recode removeCommentInSql
